% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drake.formula}
\alias{drake.formula}
\title{Estimate m using Drake method}
\usage{
drake.formula(data, e = 1, w = 1, poisson = 0, lag = 0, death = 0)
}
\arguments{
\item{data}{a vector of integer-valued colony counts.}

\item{e}{plating efficiency; a positive number not bigger than 1.}

\item{w}{mutant relative fitness; a positive number.}

\item{poisson}{average number of residual Poisson-distributed mutations on the plate; a non-negative number.}

\item{lag}{phenotypic lag (\strong{only non-stochastic}); a non-negative number.}

\item{death}{death probability of wild-type and mutant cells; a non-negative number smaller than 0.5; relative (d)eath rate
and death (p)robability are connected by the relation d = p/(1-p); p = d/(1+d).}
}
\value{
a single non-negative value of m.
}
\description{
The basic formula comes from the Drake paper. Corrections for partial plating,
phenotypic delay (\strong{only non-stochastic}), residual mutations, and differential growth were taken from other Angerer and Koch.
Correction for mutant death is based on the Angerer's rationale regarding phenotypic lag, with Newcombe correction
for extra cellular divisions.
}
\examples{
drake.formula(c(67, 12, 112, 24, 2151, 159, 102, 60, 32, 26))
}
\references{
Newcombe HB. Delayed Phenotypic Expression of Spontaneous Mutations in Escherichia Coli. Genetics.
1948;33: 447–476. doi:10.1093/genetics/33.5.447

Koch AL. Mutation and growth rates from Luria-Delbrück fluctuation tests. Mutat Res - Fundam Mol Mech Mutagen.
1982;95: 129–143. doi:10.1016/0027-5107(82)90252-4

Drake JW. A constant rate of spontaneous mutation in DNA-based microbes. Proc Natl Acad Sci U S A.
1991;88: 7160–7164. doi:10.1073/pnas.88.16.7160

Angerer WP. A note on the evaluation of fluctuation experiments. Mutat Res - Fundam Mol Mech Mutagen.
2001;479: 207–224. doi:10.1016/S0027-5107(01)00203-2
}
