% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{jones.median}
\alias{jones.median}
\title{Estimate m using Jones method of the median}
\usage{
jones.median(data, eff = 1)
}
\arguments{
\item{data}{a vector of integer-valued colony counts.}

\item{eff}{plating efficiency; a positive number not bigger than 1.}
}
\value{
a single non-negative value of m.
}
\description{
The algorithm comes from rSalvador (jones.median.plating). It is included for the sake of
self-containedness.
}
\examples{
jones.median(c(67, 12, 112, 24, 2151, 159, 102, 60, 32, 26))
}
\references{
Zheng Q. rSalvador: An R package for the fluctuation experiment. G3 Genes, Genomes, Genet.
2017;7: 3849–3856. doi:10.1534/g3.117.300120
}
