% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{koch.quartiles}
\alias{koch.quartiles}
\title{Estimate m using Koch method of quartiles}
\usage{
koch.quartiles(data, lag = 0)
}
\arguments{
\item{data}{a vector of integer-valued colony counts.}

\item{lag}{phenotypic lag; a non-negative number.}
}
\value{
a vector of length 4 containing mean m as well as first, second and third quartile.
}
\description{
The formula comes from Rosche and Foster (2000).
}
\examples{
koch.quartiles(c(67, 12, 112, 24, 2151, 159, 102, 60, 32, 26))
}
\references{
Rosche WA, Foster PL. Determining Mutation Rates in Bacterial Populations. Methods.
2000;20: 4–17. doi:10.1006/meth.1999.0901
}
