% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lrt.rate}
\alias{lrt.rate}
\title{Calculate p-values using Likelihood Ratio Test using pairs of colony counts on selective and non-selective medium}
\usage{
lrt.rate(
  data1,
  data2,
  e1 = 1,
  e2 = 1,
  w1 = 1,
  w2 = 1,
  lag1 = 0,
  lag2 = 0,
  poisson_rate1 = 0,
  poisson_rate2 = 0,
  death1 = 0,
  death2 = 0,
  inoculum1 = 0,
  inoculum2 = 0,
  Nt1,
  Nt2,
  verbose = FALSE
)
}
\arguments{
\item{data1}{a vector of integer-valued colony counts for strain 1.}

\item{data2}{a vector of integer-valued colony counts for strain 2.}

\item{e1}{plating efficiency for strain 1; a positive number not bigger than 1.}

\item{e2}{plating efficiency for strain 2; a positive number not bigger than 1.}

\item{w1}{mutant relative fitness for strain 1; a positive number.}

\item{w2}{mutant relative fitness for strain 2; a positive number.}

\item{lag1}{phenotypic lag for strain 1; a non-negative number.}

\item{lag2}{phenotypic lag for strain 2; a non-negative number.}

\item{poisson_rate1}{average mutation rate of Poisson-distributed mutations on the plate for strain 1; a non-negative number.}

\item{poisson_rate2}{average mutation rate of Poisson-distributed mutations on the plate for strain 2; a non-negative number.}

\item{death1}{death probability of wild-type and mutant cells for strain 1; a non-negative number smaller than 0.5;
relative (d)eath rate and death (p)robability are connected by the relation d = p/(1-p); p = d/(1+d).}

\item{death2}{death probability of wild-type and mutant cells for strain 2; a non-negative number smaller than 0.5;
relative (d)eath rate and death (p)robability are connected by the relation d = p/(1-p); p = d/(1+d).}

\item{inoculum1}{number of cells in the inoculum for strain 1; a non-negative integer}

\item{inoculum2}{number of cells in the inoculum for strain 2; a non-negative integer}

\item{Nt1}{a vector of integer-valued culture sizes for strain 1; must be of the same length as data.}

\item{Nt2}{a vector of integer-valued culture sizes for strain 2; must be of the same length as data.}

\item{verbose}{if TRUE, mlemur will print messages to the console.}
}
\value{
a single value of p-value between 0 and 1.
}
\description{
This function calculates LRT-based p-values to assess the statistical significance of the
differences between two mutation rates (1 and 2). The algorithm uses pairs of colony
counts on selective and non-selective medium for each test tube. It uses a hybrid Newton-bisection
algorithm as well as arbitrary-precision arithmetic if necessary for better stability.
}
\examples{
lrt.rate(data1 = c(14, 20, 59, 46, 77, 38, 86, 17, 13, 25),
Nt1 = c(1030417500, 635899434, 721044961, 1496257034, 1276590167,
1977307240, 954768054, 1192512084, 632462708, 748020573), data2 = c(387, 630,
140, 50, 1187, 18, 161, 30, 32, 61), e2 = 0.5, Nt2 = c(3543784628,
2384685853, 4744886069, 2758252418, 4388557268, 1098581801, 4363192697,
1314405541, 1677598351, 3420509139))
}
\references{
Zheng Q. Statistical and algorithmic methods for fluctuation analysis with SALVADOR as an implementation. Math Biosci.
2002;176: 237–252. doi:10.1016/S0025-5564(02)00087-1

Zheng Q. New algorithms for Luria-Delbrück fluctuation analysis. Math Biosci.
2005;196: 198–214. doi:10.1016/j.mbs.2005.03.011

Zheng Q. On Bartlett’s formulation of the Luria-Delbrück mutation model. Math Biosci.
2008;215: 48–54. doi:10.1016/j.mbs.2008.05.005

Zheng Q. Comparing mutation rates under the Luria–Delbrück protocol. Genetica.
2016;144: 351–359. doi:10.1007/s10709-016-9904-3

Zheng Q. rSalvador: An R package for the fluctuation experiment. G3 Genes, Genomes, Genet.
2017;7: 3849–3856. doi:10.1534/g3.117.300120
}
