% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{luria.delbruck.mean}
\alias{luria.delbruck.mean}
\title{Estimate m using Luria-Delbruck method of the mean}
\usage{
luria.delbruck.mean(data)
}
\arguments{
\item{data}{a vector of integer-valued colony counts.}
}
\value{
a single non-negative value of m.
}
\description{
The formula comes from Rosche and Foster (2000).
}
\examples{
luria.delbruck.mean(c(67, 12, 112, 24, 2151, 159, 102, 60, 32, 26))
}
\references{
Rosche WA, Foster PL. Determining Mutation Rates in Bacterial Populations. Methods.
2000;20: 4–17. doi:10.1006/meth.1999.0901
}
