% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mle.fold}
\alias{mle.fold}
\title{Calculate profile likelihood confidence intervals for an arbitrary function of mutation rates}
\usage{
mle.fold(
  data,
  e = NULL,
  w = NULL,
  lag = NULL,
  poisson = NULL,
  death = NULL,
  phi = NULL,
  cv = NULL,
  Nt = NULL,
  fun = "fold X1/X2",
  ci.level = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{data}{a list of length 2 to 6 containing numeric vectors with colony counts.}

\item{e}{a list or vector of length 2 to 6 (same as data) containing respective values of plating efficiency.}

\item{w}{a list or vector of length 2 to 6 (same as data) containing respective values of relative mutant fitness.}

\item{lag}{a list or vector of length 2 to 6 (same as data) containing respective values of phenotypic lag.}

\item{poisson}{a list or vector of length 2 to 6 (same as data) containing respective values of the number of residual mutations.}

\item{death}{a list or vector of length 2 to 6 (same as data) containing respective values of relative death rates.}

\item{phi}{a list or vector of length 2 to 6 (same as data) containing respective values of inoculum sizes as a fraction of total culture sizes.}

\item{cv}{a list or vector of length 2 to 6 (same as data) containing respective values of coefficient of variation.}

\item{Nt}{a list or vector of length 2 to 6 (same as data) containing respective values of average number of cells in culture.}

\item{fun}{describes the function of mutation rates for which confidence intervals should be calculated. Must be a character
vector of length 1 containing either one of default options ("fold X1/X2", "subtraction X1-X2",
"double fold (X1/X2)/(X3/X4)", "background subtraction fold (X1-X2)/(X3-X2)"), "additivity (X1+X2)/X3", or a user-defined
function containing phrases X1, X2, X3, X4, X5, X6, which denote mutation rates for Strain 1, Strain 2, ... .
Only addition +, subtraction -, multiplication *, and division /, are currently supported.}

\item{ci.level}{confidence interval size, default 0.95.}

\item{verbose}{if TRUE, mlemur will print messages to the console.}
}
\value{
a list containing two elements: "fun", a vector of length 3 containg MLE of an arbitrary
function of mutation rates as well as lower and upper profile likelihood confidence limits;
"rates", a vector containing MLEs of rates used to calculate the value of fun.
}
\description{
Inspired by Zheng 2021, this function calculates MLE and profile likelihood confidence intervals
of an arbitrary function of data, such as fold (X1/X2), subtraction (X1-X2), fold with background
subtraction ((X1-X2)/(X3-X2)), test of additivity ((X1+X2)/X3), or a user-defined function
utilising basic mathematical operations: addition, subtraction, multiplication or division. Up to 6 datasets can be used.
}
\examples{
mle.fold(list(c(26, 9, 16, 34, 15, 25, 77, 13, 14, 19), c(67, 12, 112, 24,
2151, 159, 102, 60, 32, 26)), e = c(0.5, 1))
mle.fold(list(c(33,17,15), c(1,4,10), c(45,86,156)), fun="X1*X2/X3")
}
\references{
Venzon DJ, Moolgavkar SH. A Method for Computing Profile-Likelihood-Based Confidence Intervals. Appl Stat.
1988;37: 87. doi:10.2307/2347496

Zheng Q. New approaches to mutation rate fold change in Luria–Delbrück fluctuation experiments. Math Biosci.
2021;335: 108572. doi:10.1016/j.mbs.2021.108572
}
