\name{mlemur-package}
\alias{mlemur-package}
\docType{package}
\title{mlemur: MLE Mutation Rate calculator}
\description{
  A tool for calculating maximum likelihood-based mutation rates,
  confidence intervals, and P values. Offers graphical user interface built with shiny.
}
\details{
  Based on rSalvador by Qi Zheng <https://github.com/eeeeeric/rSalvador>; GPL 2.0
}
\author{
Krystian Lazowski, klazowski@ibb.waw.pl

contributor: Qi Zheng

Maintainer: Krystian Lazowski <klazowski@ibb.waw.pl>
}
