% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlemurGUI.R
\name{mlemur}
\alias{mlemur}
\title{Run mlemur in a browser window}
\usage{
mlemur(options = list(display.mode = "normal", launch.browser = T))
}
\arguments{
\item{options}{arguments passed to shiny::shinyApp(options=...)}
}
\description{
This function initializes mlemur in graphical mode.
}
