% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{power.est}
\alias{power.est}
\title{Calculate power of likelihood ratio test}
\usage{
power.est(
  n1 = 30,
  n2 = NULL,
  rate1 = 1e-09,
  rate2 = 2e-09,
  Nt1 = 1e+09,
  Nt2 = 1e+09,
  e1 = 1,
  e2 = 1,
  w1 = 1,
  w2 = 1,
  lag1 = 0,
  lag2 = 0,
  death1 = 0,
  death2 = 0,
  phi1 = 0,
  phi2 = 0,
  cv1 = 0,
  cv2 = 0,
  poisson1 = 0,
  poisson2 = 0,
  verbose = FALSE
)
}
\arguments{
\item{n1}{A vector containing sample size of the first strain. If longer than one, powr will be calculated for each
pair of n1 and n2.}

\item{n2}{A vector containing sample size of the second strain. If NULL, it will be taken as equal to n1.}

\item{rate1}{Mutation rate of the first strain.}

\item{rate2}{Mutation rate of the second strain.}

\item{Nt1}{Average culture size of the first strain.}

\item{Nt2}{Average culture size of the second strain.}

\item{e1}{Plating efficiency of the first strain.}

\item{e2}{Plating efficiency of the second strain.}

\item{w1}{Relative mutant fitness of the first strain.}

\item{w2}{Relative mutant fitness of the second strain.}

\item{lag1}{Phenotypic lag of the first strain.}

\item{lag2}{Phenotypic lag of the second strain.}

\item{death1}{Relative death rate of the first strain.}

\item{death2}{Relative death rate of the second strain.}

\item{phi1}{Size of inoculum in proportion to the total culture size of the first strain.}

\item{phi2}{Size of inoculum in proportion to the total culture size of the second strain.}

\item{cv1}{Coefficient of variation of the culture sizes of the first strain.}

\item{cv2}{Coefficient of variation of the culture sizes of the second strain.}

\item{poisson1}{Average number of residual Poisson-distributed mutations on the plate of the first strain.}

\item{poisson2}{Average number of residual Poisson-distributed mutations on the plate of the second strain.}

\item{verbose}{if TRUE, mlemur will print messages to the console.}
}
\value{
A vector of length 1 or a matrix of size length(n1) x length(n2) containing values of the power.
}
\description{
This function calculates theoretical power of the likelihood ratio test given prescribed mutation rates,
final culture sizes, and sample sizes.
}
\examples{
power.est(n1=15, n2 = 10, rate1 = 1e-9, rate2 = 2e-9, Nt1 = 1e9, Nt2 = 5e8)
power.est(n1=c(10, 20, 30), rate1 = 1e-9, rate2 = 5e-9, Nt1 = 1e9, Nt2 = 1e9, e1 = 1, e2 = 0.1)
}
\references{
Gudicha DW., Schmittmann VD. and Vermunt JK. Statistical power of likelihood ratio and Wald tests in latent class models with covariates. Behav. Res. Methods
2017;49: 1824–1837. doi:10.3758/s13428-016-0825-y

Self SG., Mauritsen RH. and Ohara J.Power Calculations for Likelihood Ratio Tests in Generalized Linear Models. Biometrics
1992;48: 31. doi:10.2307/2532736
}
