% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample.size}
\alias{sample.size}
\title{Determine sample size to achieve prescribed power of the likelihood ratio test}
\usage{
sample.size(
  power = 0.8,
  rate1 = 1e-09,
  rate2 = 2e-09,
  Nt1 = 1e+09,
  Nt2 = 1e+09,
  e1 = 1,
  e2 = 1,
  w1 = 1,
  w2 = 1,
  lag1 = 0,
  lag2 = 0,
  death1 = 0,
  death2 = 0,
  phi1 = 0,
  phi2 = 0,
  cv1 = 0,
  cv2 = 0,
  poisson1 = 0,
  poisson2 = 0,
  verbose = FALSE
)
}
\arguments{
\item{power}{A vector containing desired statistical power. If longer than one, sample size
will be calculated for each value of power.}

\item{rate1}{Mutation rate of the first strain.}

\item{rate2}{Mutation rate of the second strain.}

\item{Nt1}{Average culture size of the first strain.}

\item{Nt2}{Average culture size of the second strain.}

\item{e1}{Plating efficiency of the first strain.}

\item{e2}{Plating efficiency of the second strain.}

\item{w1}{Relative mutant fitness of the first strain.}

\item{w2}{Relative mutant fitness of the second strain.}

\item{lag1}{Phenotypic lag of the first strain.}

\item{lag2}{Phenotypic lag of the second strain.}

\item{death1}{Relative death rate of the first strain.}

\item{death2}{Relative death rate of the second strain.}

\item{phi1}{Size of inoculum in proportion to the total culture size of the first strain.}

\item{phi2}{Size of inoculum in proportion to the total culture size of the second strain.}

\item{cv1}{Coefficient of variation of the culture sizes of the first strain.}

\item{cv2}{Coefficient of variation of the culture sizes of the second strain.}

\item{poisson1}{Average number of residual Poisson-distributed mutations on the plate of the first strain.}

\item{poisson2}{Average number of residual Poisson-distributed mutations on the plate of the second strain.}

\item{verbose}{if TRUE, mlemur will print messages to the console.}
}
\value{
A vector of length length(power) containing sample sizes (n1=n2).
}
\description{
This function calculates the sample size n1=n2 to achieve prescribed power
in the likelihood ratio test given prescribed mutation rates and final culture sizes.
}
\examples{
sample.size(power = 0.8, rate1 = 1e-9, rate2 = 2e-9, Nt1 = 1e9, Nt2 = 5e8)
sample.size(power = c(0.2, 0.5, 0.9), rate1 = 1e-9, rate2 = 5e-9, Nt1 = 1e9,
Nt2 = 3e8, e1 = 1, e2 = 0.1)
}
\references{
Gudicha DW., Schmittmann VD. and Vermunt JK. Statistical power of likelihood ratio and Wald tests in latent class models with covariates. Behav. Res. Methods
2017;49: 1824–1837. doi:10.3758/s13428-016-0825-y

Self SG., Mauritsen RH. and Ohara J.Power Calculations for Likelihood Ratio Tests in Generalized Linear Models. Biometrics
1992;48: 31. doi:10.2307/2532736
}
